Cache the dbus interface for org.kde.nepomuk.DataManagement
Regularly breaking into nepomukfeeder with gdb showed that it was always
always in the code that connects via DBus to that service, since each
and every job had to do that.

I tested what happens when killing and restarting nepomukstorage,
and QtDBus manages to reconnect just fine, with the same interface instance.
KMail locks up or crashes in Soprano::Client::SocketHandler, but that's
another story.

CCMAIL: chrigi_1@fastmail.fm, sebastian@trueg.de