Merge duplicate statements in the ResourceMerger
This happens in cases where two resources identifiy to the same resource and
their statements are resolved in the ResourceMerger. During the cardinality
checks, there exist identical duplicates (because of the merge) of certain
properties which result in cardinality errors.

Also added a unit test.

This should fix the problem with PIM Email Indexing

CCMAIL: wstephenson@kde.org