Enable Virtuoso inference on all queries.
Thanks to Virtuoso's graph groups it is fairly simple to enable
inference on all queries.

However, we need to allow clients to disable inference as there are
many situations in which inference just makes for weird results (think
property listings in the UI).
This will be a bit harder as it might require changes in Soprano.