Use text mode for writing C/C++ style lines into the file
It currently breaks the Windows build because the Windows KDE project switched to python 3.2 from python 2.7. The reason is that 'b' in the filemode does not do any encodings at all. Also, python3 strings are different to python2 strings. Python3 has the idea of "binary string".

Meanwhile, we could make something like "if sys.hexversion > 0x03000000", it is better to just drop the binary mode since it seems a valid text mode writing. Hence, we do not need "ifdef" like style for checking the python version runtime