Use black as text color in TextField and TextInput
It seems that there's not a single theme out there which properly sets
viewTextColor. The background of textedits is white, so the text has to
be black. On Air(-mobile) this is not a problem, since the text is black
anyway, on dark themes (which set the text color to white), text simply
isn't visible.

We'll have to look into how themes implement viewtextcolor, should
probably fix the themes instead of the components, but this makes it
work until we do.