Avoid using the query service to run queries in the kioslave
Avoid creating a new thread in the query service, and parsing the
results over dbus. Instead just run the query manually in the kioslave.

We can do this cause kioslaves run in a separate process and are allowed
to be blocking.

Cherry-picked from master 6e51c4136627f0721783595a9ae5468b3718c1d9