Improve response and behavior in ScrollBar
When contentHeight and contentY change at the same time, we got to
some weird states where the scrollbar would trigger a change in the
Flickable positioning that wasn't triggered by the user, this should
be fixed now.

Also don't use the handle position to move the Flickable scrolling
if the scrollbar is disabled or non-interactive. It shouldn't be needed
now but it helps keeping the complexity to a minimum.

REVIEW: 105808