auto-detect encoding of man page source
as man page files do not define in which encoding they are written,
the man-db source of the man commandline tools has a hardcoded list
of languages and their encodings (see its encodings.c file)
As this might change in the future (e.g. distributions move to UTF-8)
the hardcoded values might be wrong.
Therefore I now use KDEs encoding auto-detection mechanism and convert the pages
source always to UTF-8 before processing.

BUG: 141340
FIXED-IN: 4.9.1