Containments expose an AppletContainer
This adds a class to manipulate applets in more detail in context of
containments. The type will only be available from Containments, not
Applets of PopupApplets.

This binds the necessary functionality of Applets to build fully
features containments. The code has been moved from MobileComponents,
untangling another piece there and putting the functionality where it
belongs, into the specific component type bindings.
The code has been used for some time in Plasma Active, so can be
considered stable and well tested.