Nepomuk: New tags kioslave
It lists all the tags and lets the user browse their files based on
these tags. Each tag directory shows the files tagged with the
corresponding tag, along with any additional tags that may be used to
filter the search results.

This kio slave does not support creation/deletion/renaming or any other
operation on the tags.