dirty version to pass QActions to QML runtime
This commit contains a few tries to pass QList<QAction*> and
QDeclarativeProperty<Action> across to the QML runtime, both fail (only
one item gets shown, and it's empty).

The working solution is to pass a list of strings with the action texts
to the ListView as model, and set the actions through a QAction*
toolAction(QString) from the containment. This lets us access text, icon
and trigger() of the action.