Tags kioslave: Disable copy from file
Disabling copies for now since allowing copies means allowing move
operations.  Moving is implemented in kio as copy + delete. When moving
files from file:/ to tags:/ the tags copy operation (this function) is
called, and then the file:/ delete opteration is called, therby deleting
the file.

This needs to be fixed in kio, to not allow moving or automatically
convert a move to a copy operation.

Since we do not want the user to ever loose their files, we're disabling
copy operations