NepomukSearch: Use proper URLs
We never used to set the UDS_URL as we loose certain amount of context
information such as browsing query results in kde applications (when you
open the file). However, setting the UDS_URL makes stuff a lot easier.
And the context information was rarely used.

We can avoid setting a unique UDS_NAME, and set it as the filename
instead, cause we have a unique UDS_URL. This has the benfit of not
showing an ugly 'nepomuk_res_234-dsafdf' name when looking at the
properties of the file.

Also cleaned up the code in uds entry generation.