NepomukSearch: Cleanup the different root queries
We no longer allow the user to change the root query in the KCM. That
open is too low level, and was just scary.

Since the user can no longer change the query, there is no point trying
to read the query from the config file. We just use the default one -
last modfified files.