NepomukSearch KIO Slave: Set a proper UDS_NAME for non local files
The Nepomuk Search KIO Slave lists all the entires which have a nie:url,
and have a corresponding kioslave for the protocol. Eg - Akonadi. In
that case the urls might or might not have a file. So, we use
Resource::genericLabel instead.

This allows emails to be listed in the search results. It is kinda cool.