Nepomuk KCM: Add a button to check for new files and changes
Add a new button in the Nepomuk KCM's Advanced Settings tab. When
clicked, the button forces the indexer to check for new files in the
folders marked for indexing. This adds a greater sense of control to the
user. A new QGroupBox was added as well, with the title "Actions". In
the future, there will be more commands like this available there.

BUG: 303658
REVIEW: 107553
GUI: New button and button group in the Nepomuk KCM