Nepomuk KCM: Do not show system links in the list of directories
Considering that Nepomuk does not follow system links, it does not make
sense for them to be shown in the list of directories. Ideally, one
should show them differently and not allow them to be checked, but that
would require a lot more work.

This is more of a temporary fix.

BUG: 287593