Nepomuk KCM: Show symlinks in "Customize Folders" as disabled items
Nepomuk currently does not follow symlinks when indexing. When marking
folders to be indexed in the KCM ("Indexing" tab, "Customize Folders"
button), the user now sees the symbolic links are there, but is not
allowed to mark one for indexing, nor its children. Symbolic links
were hidden before this.

BUG: 287593
REVIEW: 108056