NepomukSearch: Always stat the file before listing it
This removes the optimization that we use the pre-fetch the stat results
from the database. This optimization is being removed because we may
have old entries in the search results for files that no longer exist.
The filewatch doesn't always work, also removable media handling doesn't
always save it under filex.

( With 4.10 we have removed the filewatch - InvalidFileResourceCleaner )

We do not want to show file results that do not exist.

Also always fetch the nie:url instead of it being optional