NepomukSearch: Avoid using the Resource class
The Resource class loads all the properties of the resource in one go
and then registers with the ResourceWatcher to watch for changes to each
of those resources. This is not particularly fast, specially when
loading a Resource which contains a large amount of plain text (pdf
files)

In the nepomuksearch kio-slave we only require the genericLabel of the
resource, which can be inferred by trying a number of different
properties. This coupled with the earlier commits results in a massive
speed difference resulting in the nepomuksearch kio-slave becoming
almost instant.

DIGEST: Massive speed improvements in the nepomuksearch kioslave