Make the BugzillaManager constructor requiring explicit URL parameter
It doesn't feel good to use a hidden default for bugs.kde.org and
require a second setCustomBugtrackerUrl() call to set a different
bugtracker address. That even doesn't take into consideration
the fact that the curent implementation of setCustomBugtrackerUrl() is
just broken: it just change the url, while the rpc connectin is still
for the old address.

In general, I think it is better to see and use BugzillaManager as
something dedicate to one tracker once it has been created. That is
especially true if we want to make it fetching and rememering more
information from the bug tracker: product information, supported
keywords, supported seveirty values, etc. It is better to create another
instance for another bug tracker instead of reusing one existing instance.