Tags kioslave: Use the Tag class internally
We had avoided using the Tag class for listDir cause that would mean n+1
queries when listing all the tags - One for fetching all their uris and
the other to fetch each tag's properties. We could instead just do it in
1 query by fetching all the uris and identifiers in one go.

However, this approach had the drawback that we would display the
nao:identifier, whereas all applications actually show the
nao:prefLabel. Also, we could not display the correct datetime of the
tags.

This approach is cleaner and uses the ResourceManager to handle all the
caching.

BUG: 311468