Nepomuk: Remove the Controller
This is probably a controversial decission, but I do not think the
controller is required any more. It's main purpose was to inform the
user when files were being indexed, so that they could pause the
indexing if they wanted.

Since 4.10, our indexing speeds have improved dramatically. With 4.11,
the indexing speeds have improved ever more, also, we have better
scheduling in place so that the indexing happens slowly over time and
the user does not notice.

In the case when stuff (files and emails) are being indexed slowly over
time, one does not want the user to be aware of it by having an active
system tray icon.

The ability to pause/resume indexing is still going to be possible
through the KCM. It's currently hidden away a little bit, but other
commits will make it more obvious.