Nepomuk KCM: Replace the StatusWidget with a DetailsWidget
This DetailsWidget is similar to the StatusWidget except for the fact
that it only shows the number of indexed files and emails. It does not
let the user suspend or resume indexing.

Also, I've moved away from the UI file and have written the QWidget code
manually. I find it easier to understand and maintain.