Nepomuk KCM: Remove the Advanced tab
We do not want to give the users an option to configure the virtuoso
memory usage since virtuoso does not respect this setting. Therefore
there is no point in giving the users a false sense of control.

Additionally the user has no information which can help them make such a
decission.