Use Phonon::PulseSupport to prevent device detection with PulseAudio
Previously whenever a device changed phononserver would rescan the
devices and display a pointless window informing the user about the
device removal. This is no longer the desired default experience as
we now recommend PulseAudio which handles device mapping and caching
for us.

Ths is wrong for two reasons.
a) With PA scanning Alsa/OSS for devices is just wrong, because we are not
   going to use them anyway
b) The Phonon KCM with PA does not even visually present the Solid based
   device cache as PA overrides everything there.
So in short whatever the user did with the notificaton window it had no
impact on anything as the cache was not used anyway and updating the cache
was pointless for the very same reason.

Now phononserver will check on the fly whether PA is active and if so,
refuse to scan for devices. This however has the disadvantage that now
phononserver actually becomes a PA client getting continous (noop)
updates about device/stream changes as we repeatedly check activity
just in case PA explodes and we actually need to update the cache.

In the long run this ought to be a one time check though as PA
auto-restarts unless something went terribly wrong, for the time being
continous checks seem less invasive to the runtime experience.

BUG: 321036