Allow always forgetting devices (no dialog whatsoever) in phononserver
In kdedrc one can now add
> [Notification Messages]
> phonon_always_forget_devices=true
which turns off the "do you want to remove these devices" window entirely.

Along with the previous change with not processing devices when Pulse is
used this should allow the majority of users to get rid of this annoying
thing once and for all.

BUG: 239983