GPG backend now fully functional, ready for real-life testing
A new wizard now appears when user wants to create a new wallet.
This wiward will invite the user to specifiy if GPG should be used
or the classic blowfish backend. The GPG wallets use a new format
and it's contents are encrypted using GpgME++

The GPG backend will not be available if GpgME++ is not found.
For the moment, GpgME++ is part of kdepimlibs.