Make the setData() accessor take a QVariant as payload.
The implementation now specifically tries to support byte arrays
being passed in, which is what the base class speaks. To try and
retain the previous behavior anything else is attempted to be
converted to a string and failing that, silently dropped.

Backport from plasma-framework f63357bd and reviewed in 111748.

CCBUG:322567