Add retry timer to the screen saver connection logic
BUG: 254273
FIXED-IN: 4.11.2

It seems that on some systems the kwalletd comes before the screen saver
service, and so it fails to connect to it. As a result, it fails to receive
ActiveChanged signal, and wallets are never closed, even if the KCM option
for this is ticked. That is before the connection is attempted only
during startup. This commit adds a retry timer to reattempt connection.

The connection is reattempted every 10s, as long as kwalletd runs and the
screen saver service is not available.