Prevent KCM crash after a context goes bad
If AudioSetup cannot connect to the context at construction the entire tab
will simply not be added, so the KCM cannot crash on nullptr access in
that case. However when a context goes bad after initial connection the KCM
did not try to reconnect which pretty much ensures a crash on s_context=0.

Connection logic was refactored into af function and used to attempt a
reconnect. Additionally when failing to connect to the context the widget
is now disabled, so even if the reconnect fails the user can not crash
AudioSetup (also without a context the widget is useless).

BUG: 267704
FIXED-IN: 4.11.3