Port Locale systemsettings module
This is a basic port, it builds and seems to basically function.

There's still a lot of KDE4Support stuff in there, and the
calendarsystem is disabled for now, as that part of the API changed a
bit. One issue I've seen is that the Plasma clock doesn't seem to pick
up the time format change, will need further investigation, possibly a
fix in Plasma.

Still, one more step towards a working desktop. :)

Screenshot: http://i.imgur.com/jld7RgS.png

DIGEST:Port Locale Systemsettings module