Port KTimeZoned to Qt5/KF5
Originally I wanted to port KTimeZoned 1:1 to Qt5/KF5, but then I found
out that all the stuff KTZD was doing was added in QTimeZone, that
includes reading correct system files/env variables to obtain the
timezone and returning the proper system zone (KTZD did all this by
itself). It also doesn't need to parse the timezone files itself or
watch for their changes as QTimeZone objects are not stored.

So now it's just a thin module watching /etc/timezone (used by Debian-
based distros) and /etc/localtime (used by eg. Fedora or Suse, but also
by Debian in conjunction with /etc/timezone) for changes and if it
detects a change, it checks if the new timezone is really different and
if it is, it sends out a DBus signal "timeZoneChange". I changed it from
"configChanged" as I think "timeZoneChanged" makes way more sense.

Besides that it also watches the timezone database and emits signal
"timeZoneDatabaseUpdated" in case the database itself is changed.

REVIEW: 113260