[kglobalaccel] Port x11 event filter to QAbstractNativeEventFilter and XCB
The KGlobalAccelImpl for X11 is switched to inherit from QObject and
QAbstractNativeEventFilter instead of QWidget. By that it doesn't
implement the x11Event filter anymore but switches to the more generic
one. All the XLib calls are ported over to XCB, only the keysym header is
included from XLib for the key defines.

The KXErrorHandler is removed and replaced by explicit request checks for
errors.