[kglobalaccel] Adjust CMakeLists for kglobalaccel
For X11 version of kglobalaccel we need to check whether we have found
the required XCB components.

Other platform builds are not yet adjusted and most likely very broken.
Given that cmake checks for Q_WS_MAC and Q_WS_WIN it will not even
compile which looks just fine at the moment.