Replace SHA with PBKDF2-SHA512+Salt
Uses the MINOR_VERSION (which until now it was 0) to upgrade the hash
from SHA to PBKDF2-SHA512+salt.
I would have loved to completely replace it once the wallet is ported
to the new hashing but because of kwalletd code that is not possible
without a bigger rewrite.

There are 2 reasons for this patch:
1-We avoid using our own implementation of SHA
2-We use a modern hashing technique

I'm cooking more patches to use the system user password to open the
wallet, we want that password to be hashed using PBKDF2_SHA512
for security reasons.

REVIEW: 115497