Do not return -2 if sync returns 0, also do not return on success
This code in openInternal is there to handle the case where no wallet
exists but somebody has requested to open it.
In pam we want to continue opening the file once it has been created
since the whole point of pam is to never ask the user for the pass.

The code is boggus because it will return -2 on sync success, and
it will return 1 when sync returns something that is not 0 which
happens on failulre.

So after this patch openInternal will create and then open the
requested kwallet. Also this is safe since openInternal is only
called once we have a password.

CCMAIL: kde@rusu.info