Remove solid-device-automounter
This should be ported to systemd so any automount happens within
systemd user session startup and we do not depend on kded being started.

Also, The current code is not working well, it keeps mounting things
after you remove the item from the list, it mounts things twice making
it bad for encrypted disks (you get the dialog twice), the model gets
fucked up etc.

So, for the time being lets just remove it, and restore it the moment
we have a maintainer and somebody to make this rock.