REVIEW: 119498. Make Dr Konqi raise its window and not crash internally.
On Apple OS X and perhaps other platforms, Dr K was invisible to the
user because it did not raise its dialog window and it could crash if
the last backtrace item had more than one line. Also added a workaround
to avoid a crash on Apple OS X when Dr K cannot find the KCookieJar.