kio_sftp: Use the right type for timeout_sec and timeout_usec.
libssh expects the values passed to the SSH_OPTIONS_TIMEOUT and
SSH_OPTIONS_TIMEOUT_USEC to be longs, not plain ints.

On 64-bit platforms with sizeof(long) > sizeof(int), this mismatch can
be problematic and potentially result in invalid memory access that
causes the calls to ssh_options_set() to fail.

BUG:      331674
FIXED-IN: 4.14.3
REVIEW:   120900