Backporting CBC algorithm fix from frameworks/kwallet
The CBC algorithm was not enrcrypting in CBC but in ECB, despite it being
named CBC. This problem was found by Itay Duvdevani who let us know it on the
security mailing list. His mail eventually got forwarded to me and here is the
fix.  I also fixed the test program which was incorrectly checking the
expected bytes.

This commit corresponds to the following commits from:
frameworks/kwallet
        88bc1ff01e5fdf59a13fe012aa03e31e9eb8a3b1
        6e588d795e6631c3c9d84d85fd3884a159b45849